u32 -> u64

+++ crates/espresso/api/src/v2/data.rs
@@ -5,20 +5,20 @@ use serialization_api::ApiSerializations;
 pub trait DataApi: ApiSerializations {
     async fn get_namespace_proof(
         &self,
-        namespace_id: u32,
+        namespace_id: u64,
         block_height: u64,
     ) -> anyhow::Result<Self::NamespaceProof>;
 
     async fn get_namespace_proof_range(
         &self,
-        namespace_id: u32,
+        namespace_id: u64,
         from: u64,
         until: u64,
     ) -> anyhow::Result<Vec<Self::NamespaceProof>>;
 
     async fn get_incorrect_encoding_proof(
         &self,
-        namespace_id: u32,
+        namespace_id: u64,
         block_height: u64,
     ) -> anyhow::Result<Self::IncorrectEncodingProof>;
 }
